moved change function to src from cli
