+++ moonshot/integrations/cli/benchmark/datasets.py
@@ -8,7 +8,7 @@ from moonshot.api import (
     api_get_all_datasets_name,
 )
 from moonshot.integrations.cli.common.display_helper import display_view_str_format
-from moonshot.integrations.cli.utils.utils import find_keyword
+from moonshot.src.utils.find_feature import find_keyword
 
 console = Console()
