+++ moonshot/integrations/cli/benchmark/recipe.py
@@ -18,7 +18,7 @@ from moonshot.api import (
     api_update_recipe,
 )
 from moonshot.integrations.cli.common.display_helper import display_view_list_format
-from moonshot.integrations.cli.utils.utils import find_keyword
+from moonshot.src.utils.find_feature import find_keyword
 
 console = Console()
