+++ moonshot/integrations/cli/benchmark/result.py
@@ -9,7 +9,7 @@ from moonshot.integrations.cli.common.display_helper import (
     display_view_list_format,
     display_view_str_format,
 )
-from moonshot.integrations.cli.utils.utils import find_keyword
+from moonshot.src.utils.find_feature import find_keyword
 
 console = Console()
