+++ moonshot/integrations/cli/benchmark/run.py
@@ -7,7 +7,7 @@ from moonshot.integrations.cli.common.display_helper import (
     display_view_list_format,
     display_view_str_format,
 )
-from moonshot.integrations.cli.utils.utils import find_keyword
+from moonshot.src.utils.find_feature import find_keyword
 
 console = Console()
