+++ moonshot/integrations/cli/redteam/context_strategy.py
@@ -10,8 +10,8 @@ from moonshot.api import (
     api_update_context_strategy,
 )
 from moonshot.integrations.cli.active_session_cfg import active_session
-from moonshot.integrations.cli.utils.utils import find_keyword
 from moonshot.src.redteaming.session.session import Session
+from moonshot.src.utils.find_feature import find_keyword
 
 console = Console()
